+++ src/linalg/LinAlg.jl
@@ -1,6 +1,0 @@
-module LinAlg
-include("./graphmatrices.jl")
-end
-
-include("./spectral.jl")
-include("./nonbacktracking.jl")