+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 - drop lg graph format
 - add Changelog
 - remove member `vertices` from graph types
+- remove graphmatrices (LinAlg submodule)